Created a Tile class - represents an ABC letter with a score in the game.
The Tile class contains a Bag class - represents the game Tiles containter with a functionality of getting tiles and returning tiles to the bag.

Created a Word class - represents an array of tiles with a first tile position (row, col) within the board.

Created a Board class - represents the tiles board game with a functionality of checking, placing and scoring word in the board.