Created a CacheManager class - represents a words container with max words size, it uses an algorithm to handle the words.
Created LRU (Last Recently Used) class - represents a container for words that has been recently asked.

Created LFU (Last Frequantly Used) class - represents a container for words that has been frequently asked.

The LRU and LFU classes implement a CacheReplacementPolicy interface represents the words replacement policy, so that LRU and LFU are different algorithms for words replacement in the cache. (Using Strategy Design Pattern)

Created BloomFilter class - represents the BloomFilter algorithm for knowing if a word does not exists in the dictionary book or if there is a high probablity if it exists.

Created IOSearcher class with a static function called search that searchs a given word in all files.

Created Dictionaty class - contains the words files, BloomFilter of the words from the files, 2 caches (exists words and not exists words) and query and challenge methods for searching words and updating the caches.